removed unnecessary casts

+++ Engine/Light.cpp
@@ -21,21 +21,21 @@ namespace CudaTracerLib {
 		unsigned int colPos = 0, rowPos = 0;
 		float rowSum = 0.0f;
 		m_cdfRows[rowPos++] = 0;
-		for (int y = 0; y < radianceMap.m_uHeight; ++y)
+		for (unsigned int y = 0; y < radianceMap.m_uHeight; ++y)
 		{
 			float colSum = 0;
 
 			m_cdfCols[colPos++] = 0;
-			for (int x = 0; x < radianceMap.m_uWidth; ++x)
+			for (unsigned int x = 0; x < radianceMap.m_uWidth; ++x)
 			{
-				Spectrum value = radianceMap.Sample(0, x, y);
+				Spectrum value = radianceMap.Sample(0, (int)x, (int)y);
 
 				colSum += value.getLuminance();
 				m_cdfCols[colPos++] = (float)colSum;
 			}
 
 			float normalization = 1.0f / (float)colSum;
-			for (int x = 1; x < radianceMap.m_uWidth; ++x)
+			for (unsigned int x = 1; x < radianceMap.m_uWidth; ++x)
 				m_cdfCols[colPos - x - 1] *= normalization;
 			m_cdfCols[colPos - 1] = 1.0f;
 
@@ -45,7 +45,7 @@ namespace CudaTracerLib {
 			m_cdfRows[rowPos++] = (float)rowSum;
 		}
 		float normalization = 1.0f / (float)rowSum;
-		for (int y = 1; y < radianceMap.m_uHeight; ++y)
+		for (unsigned int y = 1; y < radianceMap.m_uHeight; ++y)
 			m_cdfRows[rowPos - y - 1] *= normalization;
 		m_cdfRows[rowPos - 1] = 1.0f;
 		m_normalization = 1.0f / (rowSum * (2 * PI / m_size.x) * (PI / m_size.y));
